+++ apps/backend/src/crypto/key/key.module.ts
@@ -9,7 +9,9 @@ import { CrlValidationService } from "./cert/crl-validation.service";
 import { CryptoImplementatationModule } from "./crypto-implementation/crypto-implementation.module";
 import { KeyChainEntity } from "./entities/key-chain.entity";
 import { KeyChainController } from "./key-chain.controller";
+import { KeyChainImportService } from "./key-chain-import.service";
 import { KeyChainService } from "./key-chain.service";
+import { KeyChainSigningService } from "./key-chain-signing.service";
 import { KeyRotationService } from "./key-rotation.service";
 import { KmsConfigService } from "./kms/kms-config.service";
 import { KmsProviderRegistry } from "./kms/kms-provider.registry";
@@ -30,6 +32,8 @@ export class KeyModule {
                 KmsConfigService,
                 KmsProviderRegistry,
                 CertificateBuilderService,
+                KeyChainSigningService,
+                KeyChainImportService,
                 KeyChainService,
                 KeyRotationService,
                 CertService,
